@@ -6,7 +6,7 @@ homepage: https://aeyrium.com/products/mxtracker/
 repository: https://github.com/aeyrium/cube_transition
 
 environment:
-  sdk: ">=2.3.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
