@@ -7,7 +7,7 @@ import 'package:flutter/material.dart';
 /// Used by [CubePageView.builder] and other APIs that use lazily-generated widgets.
 ///
 typedef CubeWidgetBuilder = CubeWidget Function(
-    BuildContext context, int index, double pageNotifier);
+    BuildContext context, int index, double? pageNotifier);
 
 /// This Widget has the [PageView] widget inside.
 /// It works in two modes :
@@ -16,28 +16,28 @@ typedef CubeWidgetBuilder = CubeWidget Function(
 
 class CubePageView extends StatefulWidget {
   /// Called whenever the page in the center of the viewport changes.
-  final ValueChanged<int> onPageChanged;
+  final ValueChanged<int>? onPageChanged;
 
   /// An object that can be used to control the position to which this page
   /// view is scrolled.
-  final PageController controller;
+  final PageController? controller;
 
   /// Builder to customize your items
-  final CubeWidgetBuilder itemBuilder;
+  final CubeWidgetBuilder? itemBuilder;
 
   /// The number of items you have, this is only required if you use [CubePageView.builder]
-  final int itemCount;
+  final int? itemCount;
 
   /// Widgets you want to use inside the [CubePageView], this is only required if you use [CubePageView] constructor
-  final List<Widget> children;
+  final List<Widget>? children;
 
   /// Creates a scrollable list that works page by page from an explicit [List]
   /// of widgets.
   const CubePageView({
-    Key key,
+    Key? key,
     this.onPageChanged,
     this.controller,
-    @required this.children,
+    required List<Widget> this.children,
   })  : itemBuilder = null,
         itemCount = null,
         assert(children != null),
@@ -55,9 +55,9 @@ class CubePageView extends StatefulWidget {
   /// zero and less than [itemCount].
 
   CubePageView.builder({
-    Key key,
-    @required this.itemCount,
-    @required this.itemBuilder,
+    Key? key,
+    required int this.itemCount,
+    required this.itemBuilder,
     this.onPageChanged,
     this.controller,
   })  : this.children = null,
@@ -71,25 +71,25 @@ class CubePageView extends StatefulWidget {
 
 class _CubePageViewState extends State<CubePageView> {
   final _pageNotifier = ValueNotifier(0.0);
-  PageController _pageController;
+  PageController? _pageController;
 
   void _listener() {
-    _pageNotifier.value = _pageController.page;
+    _pageNotifier.value = _pageController!.page ?? 0;
   }
 
   @override
   void initState() {
     _pageController = widget.controller ?? PageController();
-    WidgetsBinding.instance.addPostFrameCallback((_) {
-      _pageController.addListener(_listener);
+    WidgetsBinding.instance!.addPostFrameCallback((_) {
+      _pageController!.addListener(_listener);
     });
     super.initState();
   }
 
   @override
   void dispose() {
-    _pageController.removeListener(_listener);
-    _pageController.dispose();
+    _pageController!.removeListener(_listener);
+    _pageController!.dispose();
     _pageNotifier.dispose();
     super.dispose();
   }
@@ -99,18 +99,18 @@ class _CubePageViewState extends State<CubePageView> {
     return Material(
       color: Colors.transparent,
       child: Center(
-        child: ValueListenableBuilder<double>(
+        child: ValueListenableBuilder<double?>(
           valueListenable: _pageNotifier,
           builder: (_, value, child) => PageView.builder(
             controller: _pageController,
             onPageChanged: widget.onPageChanged,
             physics: const ClampingScrollPhysics(),
-            itemCount: widget.itemCount ?? widget.children.length,
+            itemCount: widget.itemCount ?? widget.children!.length,
             itemBuilder: (_, index) {
               if (widget.itemBuilder != null)
-                return widget.itemBuilder(context, index, value);
+                return widget.itemBuilder!(context, index, value);
               return CubeWidget(
-                child: widget.children[index],
+                child: widget.children![index],
                 index: index,
                 pageNotifier: value,
               );
@@ -130,27 +130,27 @@ class CubeWidget extends StatelessWidget {
   final int index;
 
   /// Page Notifier value, it comes from the [CubeWidgetBuilder]
-  final double pageNotifier;
+  final double? pageNotifier;
 
   /// Child you want to use inside the Cube
   final Widget child;
 
   const CubeWidget({
-    Key key,
-    @required this.index,
-    @required this.pageNotifier,
-    @required this.child,
+    Key? key,
+    required this.index,
+    required this.pageNotifier,
+    required this.child,
   }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    final isLeaving = (index - pageNotifier) <= 0;
-    final t = (index - pageNotifier);
-    final rotationY = lerpDouble(0, 90, t);
-    final opacity = lerpDouble(0, 1, t.abs()).clamp(0.0, 1.0);
+    final isLeaving = (index - pageNotifier!) <= 0;
+    final t = (index - pageNotifier!);
+    final rotationY = lerpDouble(0, 90, t)!;
+    final num opacity = lerpDouble(0, 1, t.abs())!.clamp(0.0, 1.0);
     final transform = Matrix4.identity();
     transform.setEntry(3, 2, 0.003);
-    transform.rotateY(-degToRad(rotationY));
+    transform.rotateY(-degToRad(rotationY) as double);
     return Transform(
       alignment: isLeaving ? Alignment.centerRight : Alignment.centerLeft,
       transform: transform,
@@ -159,7 +159,7 @@ class CubeWidget extends StatelessWidget {
           child,
           Positioned.fill(
             child: Opacity(
-              opacity: opacity,
+              opacity: opacity as double,
               child: Container(
                 child: Container(
                   color: Colors.black87,
